requirement.txt
djangorestframework
six
django-bcrypt